requirements.txt
Pillow>=9.0
# piexif (optional)
# tqdm (optional)
# webcolors (optional)
